requirements.txt
opencv-python
ultralytics
whisper
soundfile
pyttsx3 
langchain
openai-whisper 
langchain-groq
sounddevice
transformers 
torch
soundfile
ultralytics
openai-clip
numpy
pillow
timm
hf_xet
peft
datasets
bitsandbytes
accelerate
albumentations
tqdm
wandb
torchvision
ipykernel
google-search-results
scikit-learn
python-dotenv
groq
